Humans Can Play
there is enough implementation for people to play. one downside is that
there are no checks for which dice contribute to points, so people can
hold any dice as long as they didn't farkle, which is cheating.
Did some basic debugging and it seems to be playing correctly, but we
need to test it more